+++ app/javascript/vue/tasks/digitize/store/actions/actions.js
@@ -3,7 +3,6 @@ import ActionNames from './actionNames'
 import addToContainer from './AddToContainer'
 import addTypeSpecimen from './addTypeSpecimen'
 import cloneCollectingEvent from './cloneCollectingEvent'
-import createDeterminationFromParams from './createDeterminationFromParams'
 import getCOCitations from './getCOCitations'
 import getCollectingEvent from './getCollectingEvent'
 import getCollectionObject from './getCollectionObject'
@@ -11,7 +10,6 @@ import getLabels from './getLabels'
 import getNamespace from './getNamespace'
 import setTypeMaterial from './setTypeMaterial'
 import setTypeMaterialTaxonName from './setTypeMaterialTaxonName'
-import getTaxonDeterminations from './getTaxonDeterminations'
 import loadTypeSpecimens from './loadTypeSpecimens'
 import loadBiologicalAssociations from './loadBiologicalAssociations'
 import loadContainer from './loadContainer'
@@ -28,16 +26,13 @@ import removeCollectionObject from './removeCollectionObject'
 import removeContainer from './removeContainer'
 import removeContainerItem from './removeContainerItem'
 import removeDepictionsByImageId from './removeDepictionsByImageId'
-import removeTaxonDetermination from './removeTaxonDetermination'
 import removeTypeMaterial from './removeTypeMaterial'
-import resetTaxonDetermination from './resetTaxonDeterminations'
 import resetWithDefault from './resetWithDefault'
 import saveCOCitations from './saveCOCitations'
 import saveCollectingEvent from './saveCollectingEvent'
 import saveCollectionObject from './saveCollectionObject'
 import saveContainer from './saveContainer'
 import saveContainerItem from './saveContainerItem'
-import saveDeterminations from './saveDeterminations'
 import saveDigitalization from './saveDigitalization'
 import saveLabel from './saveLabel'
 import saveTypeMaterial from './saveTypeMaterial'
@@ -51,12 +46,10 @@ const ActionFunctions = {
   [ActionNames.AddToContainer]: addToContainer,
   [ActionNames.AddTypeSpecimen]: addTypeSpecimen,
   [ActionNames.CloneCollectingEvent]: cloneCollectingEvent,
-  [ActionNames.CreateDeterminationFromParams]: createDeterminationFromParams,
   [ActionNames.GetCollectingEvent]: getCollectingEvent,
   [ActionNames.GetCollectionObject]: getCollectionObject,
   [ActionNames.GetLabels]: getLabels,
   [ActionNames.GetNamespace]: getNamespace,
-  [ActionNames.GetTaxonDeterminations]: getTaxonDeterminations,
   [ActionNames.SetTypeMaterial]: setTypeMaterial,
   [ActionNames.SetTypeMaterialTaxonName]: setTypeMaterialTaxonName,
   [ActionNames.LoadTypeSpecimens]: loadTypeSpecimens,
@@ -73,9 +66,7 @@ const ActionFunctions = {
   [ActionNames.RemoveContainer]: removeContainer,
   [ActionNames.RemoveContainerItem]: removeContainerItem,
   [ActionNames.RemoveDepictionsByImageId]: removeDepictionsByImageId,
-  [ActionNames.RemoveTaxonDetermination]: removeTaxonDetermination,
   [ActionNames.RemoveTypeMaterial]: removeTypeMaterial,
-  [ActionNames.ResetTaxonDetermination]: resetTaxonDetermination,
   [ActionNames.ResetStore]: resetSource,
   [ActionNames.ResetWithDefault]: resetWithDefault,
   [ActionNames.SaveCOCitations]: saveCOCitations,
@@ -83,7 +74,6 @@ const ActionFunctions = {
   [ActionNames.SaveCollectionObject]: saveCollectionObject,
   [ActionNames.SaveContainerItem]: saveContainerItem,
   [ActionNames.SaveContainer]: saveContainer,
-  [ActionNames.SaveDeterminations]: saveDeterminations,
   [ActionNames.SaveDigitalization]: saveDigitalization,
   [ActionNames.SaveLabel]: saveLabel,
   [ActionNames.SaveTypeMaterial]: saveTypeMaterial,