+++ app/javascript/vue/tasks/digitize/store/actions/loadDigitalization.js
@@ -6,12 +6,13 @@ import {
   IDENTIFIER_LOCAL_RECORD_NUMBER,
   IDENTIFIER_LOCAL_CATALOG_NUMBER
 } from '@/constants/index.js'
-import { useIdentifierStore } from '../pinia/identifiers'
+import { useIdentifierStore, useTaxonDeterminationStore } from '../pinia'
 
 export default ({ commit, dispatch, state }, coId) =>
   new Promise((resolve, reject) => {
     const recordNumber = useIdentifierStore(IDENTIFIER_LOCAL_RECORD_NUMBER)()
     const catalogNumber = useIdentifierStore(IDENTIFIER_LOCAL_CATALOG_NUMBER)()
+    const determinationStore = useTaxonDeterminationStore()
 
     state.settings.loading = true
     dispatch(ActionNames.GetCollectionObject, coId)
@@ -25,9 +26,7 @@ export default ({ commit, dispatch, state }, coId) =>
               catalogNumber.load({
                 objectId: body.id,
                 objectType: CONTAINER
-              })
-            )
-            promises.push(
+              }),
               recordNumber.load({
                 objectId: body.id,
                 objectType: CONTAINER
@@ -39,10 +38,7 @@ export default ({ commit, dispatch, state }, coId) =>
               catalogNumber.load({
                 objectId: coId,
                 objectType: COLLECTION_OBJECT
-              })
-            )
-
-            promises.push(
+              }),
               recordNumber.load({
                 objectId: coId,
                 objectType: COLLECTION_OBJECT
@@ -55,24 +51,26 @@ export default ({ commit, dispatch, state }, coId) =>
             dispatch(
               ActionNames.GetCollectingEvent,
               coObject.collecting_event_id
-            )
-          )
-          promises.push(
+            ),
             dispatch(
               ActionNames.LoadGeoreferences,
               coObject.collecting_event_id
-            )
-          )
-          promises.push(
+            ),
             dispatch(ActionNames.GetLabels, coObject.collecting_event_id)
           )
         } else {
           dispatch(ActionNames.NewLabel)
         }
 
+        promises.push(
+          determinationStore.load({
+            objectId: coId,
+            objectType: COLLECTION_OBJECT
+          })
+        )
+
         promises.push(dispatch(ActionNames.LoadTypeSpecimens, coId))
         promises.push(dispatch(ActionNames.GetCOCitations, coId))
-        promises.push(dispatch(ActionNames.GetTaxonDeterminations, coId))
         promises.push(dispatch(ActionNames.LoadBiologicalAssociations))
 
         commit(MutationNames.AddCollectionObject, coObject)