+++ app/javascript/vue/tasks/digitize/store/actions/resetStore.js
@@ -1,5 +1,5 @@
 import { makeInitialState } from '../store.js'
-import { useIdentifierStore } from '../pinia/identifiers.js'
+import { useIdentifierStore, useTaxonDeterminationStore } from '../pinia'
 import {
   EVENT_TAXON_DETERMINATION_FORM_RESET,
   IDENTIFIER_LOCAL_CATALOG_NUMBER,
@@ -15,6 +15,7 @@ export default ({ state }) => {
   const { preferences, project_preferences } = state
   const recordNumber = useIdentifierStore(IDENTIFIER_LOCAL_RECORD_NUMBER)()
   const catalogNumber = useIdentifierStore(IDENTIFIER_LOCAL_CATALOG_NUMBER)()
+  const determinationStore = useTaxonDeterminationStore()
 
   history.replaceState(null, null, '/tasks/accessions/comprehensive')
   state = Object.assign(state, makeInitialState())
@@ -22,6 +23,7 @@ export default ({ state }) => {
   state.project_preferences = project_preferences
   recordNumber.$reset()
   catalogNumber.$reset()
+  determinationStore.$reset()
 
   resetTaxonDeterminationForm()
 }