+++ app/javascript/vue/tasks/digitize/store/mutations/TaxonDetermination/setTaxonDeterminations.js
@@ -1,3 +1,0 @@
-export default (state, value) => {
-  state.taxon_determinations = value
-}